Module Deriv.Inducing

Derivatives of the covariance matrix of inducing inputs

type upper

Representation of precomputed data for calculating the upper triangle of the derivative of the covariance matrix of inducing inputs.

val calc_shared_upper : Eval.Kernel.t -> Eval.Inducing.t -> Lacaml.D.mat * upper

calc_shared_upper kernel inducing

returns
the pair (eval, upper), where eval is the upper triangle of the covariance matrix of inducing inputs for kernel, and upper is the precomputed data needed for taking derivatives.

val calc_deriv_upper : upper -> Hyper.t -> Interfaces.symm_mat_deriv

calc_deriv_upper upper hyper

returns
the derivative of the (symmetric) covariance matrix of inducing inputs given precomputed data upper and the hyper-variable.