Module Deriv.Inputs

Derivatives of the (cross-) covariance matrix of inputs.

type diag

Representation of precomputed data for calculating the derivative of the diagonal of the covariance matrix of inputs.

type cross

Representation of precomputed data for calculating the derivative of the cross-covariance matrix between inputs and inducing inputs.

val calc_shared_diag : Eval.Kernel.t -> Eval.Inputs.t -> Lacaml.D.vec * diag

calc_shared_diag kernel inputs

returns
the pair (eval, diag), where eval is the diagonal of the covariance matrix of inputs for kernel, and diag is the precomputed data needed for taking derivatives.

val calc_shared_cross : Eval.Kernel.t -> inputs:Eval.Inputs.t -> inducing:Eval.Inducing.t -> Lacaml.D.mat * cross

calc_shared_cross kernel ~inputs ~inducing

returns
the pair (eval, cross), where eval is the cross-covariance matrix of inputs and inducing inputs for kernel, and diag is the precomputed data needed for taking derivatives.

val calc_deriv_diag : diag -> Hyper.t -> Interfaces.diag_deriv

calc_deriv_diag diag hyper

returns
the derivative of the diagonal of the covariance matrix of inputs given precomputed data diag and the hyper-variable.

val calc_deriv_cross : cross -> Hyper.t -> Interfaces.mat_deriv

calc_deriv_cross cross hyper

returns
the derivative of the cross-covariance matrix of the inputs and inducing inputs given precomputed data cross and the hyper-variable.