Module Make.FIC

module Spec = Spec

Specification of covariance function

module Inducing = FITC.Inducing

Evaluating inducing inputs

module Input = FITC.Input

Evaluating single inputs

module Inputs = FITC.Inputs

Evaluating (multiple) inputs

module Model = FITC.Model

(Untrained) model - does not require targets

module Trained = FITC.Trained

Trained model - requires targets

module Stats : sig ... end

Statistics derived from trained models

module Mean_predictor = FITC.Mean_predictor

Module for making mean predictions

module Mean = FITC.Mean

Posterior mean for a single input

module Means = FITC.Means

Posterior means for (multiple) inputs

module Co_variance_predictor = FITC.Co_variance_predictor

Module for making (co-)variance predictions

module Variance = FITC.Variance

Posterior variance for a single input

module Variances = FITC.Variances

Posterior variances for (multiple) inputs

module Covariances : sig ... end

Posterior covariances

module Sampler = FITC.Sampler

Module for sampling single points from the posterior distribution

module Cov_sampler : sig ... end

Module for sampling (multiple) points from the posterior distribution accounting for their covariance