Module Deriv.Model

(Untrained) model with derivative information

type t

Type of models with derivatives

type hyper_t

Type of models for general hyper parameters

val calc : Inputs.t -> sigma2:float -> t

calc inputs ~sigma2

returns
model with derivative information given inputs and noise level sigma2.

val update_sigma2 : t -> float -> t

update_sigma2 model sigma2

returns
model with derivative information by updating model with new noise level sigma2.

val calc_eval : t -> Eval.Model.t

calc_eval model

returns
model without derivative information given model.

val calc_log_evidence_sigma2 : t -> float

calc_log_evidence_sigma2 model

returns
the derivative of the log evidence of model with respect to the noise level (sigma2).

val prepare_hyper : t -> hyper_t

prepare_hyper model

returns
the model prepared for calculating derivatives for arbitrary hyper parameters.

val calc_log_evidence : hyper_t -> Spec.Hyper.t -> float

calc_log_evidence hyper_t hyper

returns
the derivative of the log evidence given prepared model hyper_t with respect to the hyper variable.