Module Make_FIC_deriv.Eval

Sub-modules for learning without derivatives.

module Spec = Spec.Eval

Specification of covariance function

module Inducing : sig ... end

Evaluating inducing inputs

module Input : sig ... end

Evaluating single inputs

module Inputs : sig ... end

Evaluating (multiple) inputs

module Model : sig ... end

(Untrained) model - does not require targets

module Trained : sig ... end

Trained model - requires targets

module Stats : sig ... end

Statistics derived from trained models

module Mean_predictor : sig ... end

Module for making mean predictions

module Mean : sig ... end

Posterior mean for a single input

module Means : sig ... end

Posterior means for (multiple) inputs

module Co_variance_predictor : sig ... end

Module for making (co-)variance predictions

module Variance : sig ... end

Posterior variance for a single input

module Variances : sig ... end

Posterior variances for (multiple) inputs

module Covariances : sig ... end

Posterior covariances

module Sampler : sig ... end

Module for sampling single points from the posterior distribution

module Cov_sampler : sig ... end

Module for sampling (multiple) points from the posterior distribution accounting for their covariance