Module Eval.Sampler

Module for sampling single points from the posterior distribution

type t

Type of sampler

val calc : ?predictive:bool -> Mean.t -> Variance.t -> t

calc ?predictive mean variance

returns
sampler given mean and variance. If predictive is true, the samples will be noisy.

val sample : ?rng:Gsl.Rng.t -> t -> float

sample ?rng sampler

returns
a sample from the posterior distribution given sampler and GSL random number generator rng.

parameter rng
default = GSL default

val samples : ?rng:Gsl.Rng.t -> t -> n:int -> Lacaml.D.vec

samples ?rng sampler ~n

returns
n samples from the posterior distribution given sampler.

parameter rng
default = GSL default