Module Eval.Inducing

Evaluating inducing inputs

type t

Type of inducing inputs

val choose_n_first_inputs : Spec.Kernel.t -> Spec.Inputs.t -> n_inducing:int -> Spec.Inducing.t

choose_n_first_inputs kernel inputs ~n_inducing

returns
the first n_inducing inputs in inputs as inducing points given kernel.

val choose_n_random_inputs : ?rnd_state:Core.Random.State.t -> Spec.Kernel.t -> Spec.Inputs.t -> n_inducing:int -> Spec.Inducing.t

choose_n_random_inputs ?rnd_state kernel inputs ~n_inducing

returns
n_inducing random inputs in inputs as inducing points given kernel and (optional) random state rnd_state.

parameter rnd_state
default = default used by the Random module

val calc : Spec.Kernel.t -> Spec.Inducing.t -> t

calc kernel inducing_points

returns
inducing inputs (= precomputed data) prepared using inducing_points and kernel.

val get_points : t -> Spec.Inducing.t

get_points kernel inducing

returns
inducing points associated with the prepared inducing inputs.