Module Deriv.Test

Module for testing derivative code

val check_deriv_hyper : ?eps:float -> ?tol:float -> Eval.Spec.Kernel.t -> Eval.Spec.Inducing.t -> Eval.Spec.Inputs.t -> Spec.Hyper.t -> unit

check_deriv_hyper ?eps ?tol kernel inducing_points points hyper will raise Failure if the derivative code provided in the specification of the covariance function given parameter hyper, the kernel, inducing_points and input points exceeds the tolerance tol when compared to finite differences using epsilon eps. The failure exception will contain details on which derivative matrix was incorrect and indicate the matrix element.

parameter eps
default = 1e-8

parameter tol
default = 1e-2

val self_test : ?eps:float -> ?tol:float -> Eval.Spec.Kernel.t -> Eval.Spec.Inducing.t -> Eval.Spec.Inputs.t -> sigma2:float -> targets:Lacaml.D.vec -> [ `Sigma2 | `Hyper of Spec.Hyper.t ] -> unit

self_test ?eps ?tol kernel inducing_points points ~sigma2 ~targets hyper will raise Failure if the internal derivative code for the log evidence given parameter hyper, the kernel, inducing_points, input points, noise level sigma2 and targets exceeds the tolerance tol when compared to finite differences using epsilon eps.

parameter eps
default = 1e-8

parameter tol
default = 1e-2