Module Deriv.Trained

Trained model with derivative information

type t

Type of trained models with derivatives

type hyper_t

Type of trained models for general hyper parameters

val calc : Model.t -> targets:Lacaml.D.vec -> t

calc model ~targets

returns
trained model with derivative information given the untrained model and targets.

val calc_eval : t -> Eval.Trained.t

calc_eval trained

returns
trained model without derivative information given trained.

val calc_log_evidence_sigma2 : t -> float

calc_log_evidence_sigma2 trained

returns
the derivative of the log evidence for the trained model with respect to the noise level (sigma2). This includes the contribution to the derivative by model.

val prepare_hyper : t -> hyper_t

prepare_hyper trained

returns
the trained model prepared for calculating derivatives for arbitrary hyper parameters.

val calc_log_evidence : hyper_t -> Spec.Hyper.t -> float

calc_log_evidence hyper_t hyper

returns
the derivative of the log evidence given prepared, trained model hyper_t with respect to the hyper variable.