Module Eval.Stats

Statistics derived from trained models

type t = {

n_samples : int;
(*
Number of samples used for training
*)

target_variance : float;
(*
Variance of targets
*)

sse : float;
(*
Sum of squared errors
*)

mse : float;
(*
Mean sum of squared errors
*)

rmse : float;
(*
Root mean sum of squared errors
*)

smse : float;
(*
Standardized mean squared error
*)

msll : float;
(*
Mean standardized log loss
*)

mad : float;
(*
Mean absolute deviation
*)

maxad : float;
(*
Maximum absolute deviation
*)

}

Type of full statistics

val calc_n_samples : Trained.t -> int

calc_n_samples trained

returns
number of samples used for training trained.

val calc_target_variance : Trained.t -> float

calc_target_variance trained

returns
variance of targets used for training trained.

val calc_sse : Trained.t -> float

calc_sse trained

returns
the sum of squared errors of the trained model.

val calc_mse : Trained.t -> float

calc_mse trained

returns
the mean sum of squared errors of the trained model.

val calc_rmse : Trained.t -> float

calc_sse trained

returns
the root of the mean sum of squared errors of the trained model.

val calc_smse : Trained.t -> float

calc_smse trained

returns
the standardized mean squared error of the trained model. This is equivalent to the mean squared error divided by the target variance.

val calc_msll : Trained.t -> float

calc_msll trained

returns
the mean standardized log loss. This is equivalent to subtracting the log evidence of the trained model from the log evidence of a normal distribution fit to the targets, and dividing the result by the number of samples.

val calc_mad : Trained.t -> float

calc_mad trained

returns
the mean absolute deviation of the trained model.

val calc_maxad : Trained.t -> float

calc_mad trained

returns
the maximum absolute deviation of the trained model.

val calc : Trained.t -> t

calc trained

returns
the full set of statistics associated with the trained model.