Module Spec.Inputs

Signature for evaluating multiple inputs

type t

Type of input points

val create : Input.t array -> t

create inputs

returns
inputs given an array of single inputs.

val get_n_points : t -> int

get_n_points inputs

returns
number of input points.

val choose_subset : t -> Utils.Int_vec.t -> t

choose_subset inputs indexes

returns
subset of input points from inputs having indexes.

val create_inducing : Kernel.t -> t -> Inducing.t

create_inducing kernel inputs

returns
inducing points made from inputs and given kernel.

val create_default_kernel_params : t -> n_inducing:int -> Kernel.params

create_default_kernel_params inputs ~n_inducing

returns
default kernel parameters to be used with n_inducing inducing points and inputs.

val calc_upper : Kernel.t -> t -> Lacaml.D.mat

calc_upper kernel inputs

returns
upper triangle of covariance matrix of inputs given kernel.

val calc_diag : Kernel.t -> t -> Lacaml.D.vec

calc_diag kernel inputs

returns
diagonal of covariance matrix of inputs given kernel.

val calc_cross : Kernel.t -> inputs:t -> inducing:Inducing.t -> Lacaml.D.mat

calc_cross kernel ~inputs ~inducing

returns
cross-covariance matrix of inputs (indexing rows) and inducing points (indexing columns).

val weighted_eval : Kernel.t -> inputs:t -> inducing:Inducing.t -> coeffs:Lacaml.D.vec -> Lacaml.D.vec

weighted_eval kernel ~inputs ~inducing ~coeffs

returns
vector of coeff-weighted sums of covariances between inputs and inducing inputs given kernel.