Module Interfaces.Sigs

Signatures for learning sparse Gaussian processes with inducing inputs

module type Eval = sig ... end

Modules for learning without derivatives of covariance functions.

module type Deriv = sig ... end

Modules for learning with derivatives of the log evidence (evidence maximization framework)

module type Optimizer = sig ... end

Modules for global optimization