Module Spec.Input

Signature for evaluating single inputs

type t

Type of input point

val eval : Kernel.t -> t -> Inducing.t -> Lacaml.D.vec

eval kernel input inducing

returns
(row) vector of covariance evaluations between input and inducing inputs given kernel.

val weighted_eval : Kernel.t -> t -> Inducing.t -> coeffs:Lacaml.D.vec -> float

weighted_eval kernel input inducing ~coeffs

returns
coeff-weighted sum of covariances between input and inducing inputs given kernel.

val eval_one : Kernel.t -> t -> float

eval_one kernel point

returns
variance of point given kernel.