Module Spec.Input

val get_vars : Eval.Input.t -> Var.t array

get_vars input

returns
array of all input parameters for which derivatives can be computed given input.

val get_value : Eval.Input.t -> Var.t -> float

get_value input var

returns
value of input parameter var for input.

val set_values : Eval.Input.t -> Var.t array -> Lacaml.D.vec -> Eval.Input.t

set_values input vars values

returns
input in which vars have been substituted with values position-wise.