Module Spec.Inputs

val get_vars : Eval.Inputs.t -> Var.t array

get_vars inputs

returns
array of all input parameters for which derivatives can be computed given inputs.

val get_value : Eval.Inputs.t -> Var.t -> float

get_value inputs var

returns
value of input parameter var for inputs.

val set_values : Eval.Inputs.t -> Var.t array -> Lacaml.D.vec -> Eval.Inputs.t

set_values inputs vars values

returns
inputs in which vars have been substituted with values position-wise.