Module Gpr__Block_diag

type t = private {

data : Lacaml.D.mat array;

n : int;

}

Type of block diagonal matrices

val create : Lacaml.D.mat array -> t

create mats

returns

a block diagonal matrix whose block elements are made of the matrices in mats.

val copy : t -> t

copy bm

returns

a copy of block diagonal matrix bm.

val potrf : t -> unit

potrf bm perform Cholesky factorization on block diagonal matrix bm.

val potri : t -> unit

potri bm invert block diagonal matrix bm using its already precomputed Cholesky factor.